Add support for imu/gps message streaming (#6)
* Add code to support imu/gps data messages

The bulk of this update adds helper classes and functions to work with IMU and GPS data. Both were done in this update to make sure the approach would work for more than one format.

* Add unit test coverage for message streaming

Adds unit tests for new functionality in microstrain library

* Add test commands for HW data message handling

Adds new example cases that show how to configure data streaming from the device, and collect it.